Don't defer sanitization.
We want to sanitize content as soon as possible after it is entered;
this way, there is no flash of content that immediately gets removed/
modified. Deferring with setTimeout introduced a perceptible lag,
which is unacceptable.

Assuming the input event is fired synchronously, the editable element
is now sanitized synchronously, too; unfortunately, this means any
plugin added after richMode will see the chain of events as:

    beforeClean -> afterClean -> input -> change

This is not ideal, but I can't think of anything at the moment that
would require a plugin to subscribe to the input event; chances are,
what they really want is the afterClean event. If this causes any
problems in the future, we could instead consider a setImmediate
polyfill (e.g. https://github.com/YuzuJS/setImmediate); this might be
fast enough not to cause an unwanted flash of content.