More flexible paragraph/section operations
Instead of a mess of if-else statements in the paragraph/section
operations, a new module, “converter,” now handles the conversion
between elements and their abstract representations. For paragraphs
that must be children of other elements (e.g. <li>s, and someday
tables), the Converter constructs the appropriate tree for that
element; the paragraph operations then determine if any elements in
that tree can be merged with others already in the editor.