richMode: more lenience when initializing
The correctness of the HTML in the editor at initialization has been
relaxed somewhat. Whitespace gets stripped, and <hr>s are inserted at
the start of sections if not present. Additionally, if the editor is
empty (or is emptied by, for example, stripping all the whitespace),
a section is created for you with a blank paragraph.

Downsides of the new setup: in addition to stripping whitespace, it
also strips everything else it doesn’t like. If, for example, you
forget to put a <section> but still have an <hr> and some paragraphs,
all of those will get removed — the editor will thus be blank, so a
single paragraph will be inserted as per above.