richMode/selection: normalize less often
This commit actually fixes two issues.

The first issue was that we were normalizing after every keydown, which
was causing issues when, for example, entering accented characters via
Alt shortcuts on a Mac.

The second issue involved tabbing into the editor in Firefox; the caret
is placed at the very start of the element (before even any
<section>s), and Choice was interpreting this as the selection on the
<hr>; that sucked. The <hr>s were made non-contenteditable, and Choice
was updated to ignore non-contenteditable elements.