richMode: standardize render times.
The View now sets a setImmediate whenever a keydown events occurs,
which ensures that renders occur at a predictable time; previously,
the View would set a setImmediate whenever a plugin decided it wanted
to render changes, which left some uncertainty as to the order of
events.

Additionally, events are now emitted for each Delta type; these events
are emitted before any changes to either the view or the DOM have taken
place.

The changes to how the view renders broke a few tests, specifically
those dealing with when paragraph/section operations throw. I’m not
sure how to test these anymore, so, for now, they are being skipped.