view: don’t cache queue length
This was messing with inline formatting; if the length of the queue
was initially 0, and the sync resulted in a delta being added to the
queue, the render was still terminated as the previous queue length
was being stored in the "len" variable.

For whatever reason, this was not breaking the automated tests.