choice@1.1.0
Previously, Selection.get( ) returned the same instance of
Choice.Selection each time it was called; this meant changes made to
the selection in one module would be seen in others. While there is
probably a use case for this, it seems much saner to simply return
a new instance of Choice.Selection each time Selection.get is called,
which is what this commit does thanks to the changes introduced in
Choice@1.1.0.