dom: don’t expose block elements
The idea behind exposing block elements was that plugin could add to
this list. Not only will this result in shared state (dom is not an
instance-specific module), but it results in inconsistencies; its
best to always treat LIs as blocks (visually, they are), rather than
only when one uses the list plugin.